[RFC] GPIO: Change behavior of IMASK
Do not mask ILAT with IMASK. Instead mask gpio_irq with IMASK.

This should work better with edge triggered interrupts where you want to
ack (+mask) the interrupt early in the interrupt handler so that you can
receive one (but only one) consecutive interrupt on the same pin while
still in the interrupt handler.

Not 100% about this but this seems to be the way Linux wants interrupt
controllers to behave.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>