spi: Rewrite slave to be CPOL=0 CPHA=0 compliant.
CPOL=0 CPHA=0 mode:
RX data is sampled on rising edge.
TX data is outputted on falling edge.

Between the first byte is received and the second byte should start
transmitting, we have a half clock cycle so we need to bypass the
deserializer for the last MOSI bit and grab it directly from the pin.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>