Changed license copyright
The Parallella Foundation was resolved. Too much of a headache to maintain.  Considering a permanent home in an existing foundation like eclipse, or apache, or other? For now assigning to me. History of code copyright: Adapteva-->Parallella Foundation-->Andreas Olofsson

1.) Most of code developed while Andreas Olofsson was employed at Adapteva

2.) 2016: Adapteva board resolution transfered code to Parallella Foundation under leadership of Andreas Olofsson

3.) 2016: Parallella foundation resolved and code copyright transferred to Andreas Olofsson

None of this should matter given the MIT nature of the license....